+++ cleo/formatters/__init__.py
@@ -1,11 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .output_formatter import OutputFormatter
-from .output_formatter_style import OutputFormatterStyle
-from .output_formatter_style_stack import OutputFormatterStyleStack
-
-__all__ = [
-    'OutputFormatter',
-    'OutputFormatterStyle',
-    'OutputFormatterStyleStack'
-]
+from .formatter import Formatter