+++ tests/commands/completion/fixtures/command_with_colons.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+
+from cleo import Command
+
+
+class CommandWithColons(Command):
+    """
+    Test.
+    
+    command:with:colons
+        { --goodbye }
+    """