+++ tests/commands/completion/fixtures/hello_command.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+
+from cleo import Command
+
+
+class HelloCommand(Command):
+    """
+    Complete me please.
+
+    hello
+        { --dangerous-option= : This $hould be `escaped`. }
+        { --option-without-description }
+    """