+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+- Added a new `completions` command to generate autocompletion scripts.
+- Added support for command signature inheritance.
+- Added a new `spin()` helper to display a spinner.
+
+### Changed
+
+- Removed the `_completion` command.
+- Removes ability to choose when a command anme is ambiguous.
+
 ## [0.5.0] - 2016-09-21
 
 ### Added
@@ -58,7 +71,7 @@ This is a major release with some API changes.
 
 
 
-[Unreleased]: https://github.com/sdispater/cleo/compare/0.5.0...master
+[Unreleased]: https://github.com/sdispater/cleo/compare/master...develop
 [0.5.0]: https://github.com/sdispater/cleo/releases/tag/0.5.0
 [0.4.1]: https://github.com/sdispater/cleo/releases/tag/0.4.1
 [0.4]: https://github.com/sdispater/cleo/releases/tag/0.4